APITask Solution.
